requirements.txt
aif360==0.6.1
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
sklearn-compat==0.1.5
threadpoolctl==3.6.0
